Create dedicated files for each type category
