+++ js/typescript-library/src/__tests__/dependencies/computePeriodDependencies.test.ts
@@ -1,7 +1,8 @@
 import {TableField} from "../../field";
 import * as dependencies from "../../dependencies";
 import {Month, Quarter, Semester, Year} from "../../columnsets";
-import {Period} from "../../types";
+
+import {Period} from "../../types/columnsets";
 
 afterEach(() => {
   jest.restoreAllMocks();
