+++ js/typescript-library/src/pivotTableQuery.ts
@@ -1,6 +1,5 @@
 import {Query} from "./query"
-
-import {Field} from "./types";
+import {Field} from "./types/field";
 
 export interface PivotTableQuery {
   query: Query
