+++ js/typescript-library/src/types/columnsets.ts
@@ -1,0 +1,12 @@
+export interface ColumnSet {
+  readonly class: string
+  readonly key: string
+}
+
+export interface Period {
+  readonly class: string,
+}
+
+export enum ColumnSetKey {
+  BUCKET = "BUCKET",
+}
