+++ js/typescript-library/src/types/field.ts
@@ -1,0 +1,11 @@
+export interface Field {
+  readonly class: string
+
+  divide(other: Field): Field
+
+  minus(other: Field): Field
+
+  multiply(other: Field): Field
+
+  plus(other: Field): Field
+}
